tools/jira/jira_upload_attachment/requirements.txt
jira==3.5.2
python-dotenv==1.0.0